==> Scripts/09_Plot_Preg_Changes.Rmd <==
---
title: "09_Plot_Preg_Changes"
author: "Calen P. Ryan"
date: "`r format(Sys.time(), '%d %B, %Y')`"
output: 
  html_document:
  code_folding: hide

---

Started around December 7th, 2018

2005: 
 * Take the pregord from repro.final.nsf
 * Date:
  - If !is.na(yearprg, monthprg, dayprg)  
    - use the blood sample date as pregterm.date
    - This will happen for all nulliparous women.
    
NSF:
 * Take the max.all.preg
 * If missing date...


# Setup: Options 
```{r}
knitr::opts_chunk$set(tidy = TRUE)
```


# Setup: Packages - messages hidden
```{r message=F, warning=F, include=FALSE, paged.print=FALSE}
library(tidyverse) # The tidyverse, baby
library(haven) # Used to load different datatypes, like .dat
library(readr) # Also used to read different data types
library(lubridate) # Working with dates
library(expss) # 
library(sjlabelled)
library(vcd)## used for estimating goodness of fit for poisson distribution
library(here)

```

# Load the data for 2005. This is from 2007, but corrected. 
```{r}
reprostat05_w_pregord <-read_csv(here::here("Output/Data", "reprostat05_w_pregord.csv"))
```

# Load the data for 2009 and past.
# Pull those with placenta and the max/highest preg number in NSF study

## Load 2014 women 
...with placenta data, plus those without but the max pregnancy in the NSF.

```{r, message = FALSE}
placenta.women <-read_csv(here::here("Output/Data","placenta_women.csv"))
all.nsf.fem.no0preg.maxNSF <-read_csv(here::here("Output/Data","all_nsf_fem_no0preg_maxNSF.csv"))
```

```{r}
nsf.picks <-c(as.character(placenta.women$uncchdid), as.character(all.nsf.fem.no0preg.maxNSF$uncchdid))

nsf.picks <-nsf.picks[!duplicated(nsf.picks)]
```


```{r}
reprostat05.nsf.selected <-reprostat05_w_pregord %>% 
  filter(uncchdid %in% as.character(all.nsf.fem.no0preg.maxNSF$uncchdid)) %>% 
  filter(reprostat == "pregnant" | reprostat == "breastfeeding" | uncchdid %in% as.character(placenta.women$uncchdid))
```

```{r message=FALSE, warning=FALSE}
nsf.long.max <-read_csv(here::here("Output/Data","preg_all_57914_long.csv"))
names(nsf.long.max)
dim(nsf.long.max)
```

```{r message=FALSE, warning=FALSE}
# Find the max "new.preg"
nsf.single.max <-nsf.long.max %>% 
  group_by(uncchdid) %>%
  mutate(new.preg.term.date = ymd(paste(yearprg, monthprg, dayprg, sep = "-"))) %>% 
  select(-dayblood, -monthblood, -yearblood, -monthprg, -dayprg, -yearprg) %>% 
  slice(which.max(new.preg)) 

dim(nsf.single.max)
```


```{r}
# Pull the ones in the final subset (repro_final_nsf_whole.csv)
not.it <-intersect(names(reprostat05.nsf.selected), names(nsf.single.max))
not.it
names(reprostat05.nsf.selected)
names(nsf.single.max)
```

```{r}

nsf.single.max.clean <-nsf.single.max %>% 
  select(-not.it, uncchdid)

nsf.selected.w.max <-left_join(reprostat05.nsf.selected, nsf.single.max.clean, by = "uncchdid")

names(nsf.selected.w.max)

```

```{r}
pregord.long <-nsf.selected.w.max %>% 
  select(uncchdid, reprostat, pregord_2005, term.or.draw.date, new.preg, new.preg.term.date) %>% 
  gather("sample", "date", term.or.draw.date, new.preg.term.date) %>% 
  gather("time", "pregs", pregord_2005, new.preg) %>% 
  arrange(uncchdid)
  
pregord.long.clean <-pregord.long[pregord.long$sample == "term.or.draw.date" 
       & pregord.long$time == "pregord_2005" |
         pregord.long$sample == "new.preg.term.date"
       & pregord.long$time == "new.preg",]
 
pregord.long.clean
```

```{r}
pregord.long.clean$time <- factor(pregord.long.clean$time, levels = c("pregord_2005", "new.preg"))

  ggplot(pregord.long.clean, aes(x = time, y = pregs, color = reprostat)) +
  geom_line(aes(group = uncchdid), size=2, alpha = 0.5)
    
  
pregord.long.clean %>% 
  arrange(uncchdid) %>% 
  ggplot(., aes(x = date, y = pregs, color = reprostat, group = uncchdid)) +
  geom_line(aes(group = uncchdid), size=1, alpha = 0.3)+
  scale_x_date(date_breaks = "2 year", date_labels =  "%Y" )+
  theme_bw()+
  theme(legend.position = c(0.1, 0.8))
 
# Not sure how to get the x-axis a bit smaller so that one observations isn't crunched in, but I did it. 
```

```{r}
table(nsf.selected.w.max$reprostat)
```

```{r}
nsf.selected.w.max %>% 
  group_by(uncchdid) %>% 
  ggplot(aes(x = max.all.preg)) +
  geom_histogram()
```

```{r}
nsf.selected.w.max %>% 
  mutate(diff.preg = max.all.preg - pregord_2005) %>% 
  ggplot(aes(x = diff.preg, fill = reprostat)) +
  geom_bar(width = 0.4, alpha = 0.8) +
#  scale_fill_manual(values = c("skyblue", "pink", "purple", "red"))+
  theme_bw() +
  theme(legend.position = c(0.9, 0.8))
```

```{r}

# write_csv(pregord.long.clean, here::here("Output/Data/pregord.long.clean.csv"))
```
